Move Item.ToSynQPayload to SynQProductPayload
Item -> Product good
